add summary to data_date_columns.
