+++ data/timeseries.json
@@ -7,9 +7,9 @@
     "is_dirty": false
   },
   "model_git_info": {
-    "sha": "72e08ed123c66587ccfa20cdadff42d4e6051ad9",
+    "sha": "74f07f959e54d60098afe94b3fea1a209de637de",
     "branch": "tgb-wide-dates",
     "is_dirty": true
   },
-  "updated_at": "2020-07-26T16:04:09.004446"
+  "updated_at": "2020-07-27T07:40:57.774533"
 }
